MSVC warnings clean up.
Chromium is using warnings as errors w/their libvpx builds; this fixes
the chromium builds of libvpx.

+++ vp9/common/vp9_convolve.c
@@ -145,7 +145,7 @@ static const interp_kernel *get_filter_base(const int16_t *filter) {
 }
 
 static int get_filter_offset(const int16_t *f, const interp_kernel *base) {
-  return (const interp_kernel *)(intptr_t)f - base;
+  return (int)((const interp_kernel *)(intptr_t)f - base);
 }
 
 void vp9_convolve8_horiz_c(const uint8_t *src, ptrdiff_t src_stride,
